feat(16-02): extract useIsMobile + add useDashboardStatus polling hook
- apps/web/src/hooks/useIsMobile.ts: verbatim extraction of __root.tsx:20-32 per D-13. Default breakpoint 640px (Tailwind sm).
- apps/web/src/hooks/useDashboardStatus.ts: TanStack Query polling hook with queryKey ['dashboard-status'], staleTime 10_000, refetchInterval 30_000 per D-07/08/09. Param shape `string | null | undefined` accepts useSchoolContext schoolId verbatim (WARNING W3 fix); null/undefined disable the query.
- apps/web/src/routes/__root.tsx: removed local useIsMobile, now imports from @/hooks/useIsMobile.

10/10 unit tests green (4 useIsMobile + 6 useDashboardStatus including the null/undefined gate).
